Prototyped Subsystem Coordinator and created utils.time
utils.time (ToasterTimer interface, WPITimer and JavaTimer) has been
created. The basic foundation of the Subsystem Coordinator has been
laid. The current functionality is that it maps variables to HashMaps so
that these variables can be accessed whenever needed (e.x for the drive
to check if another system is ready to shoot or whatever). There is also
a timer in the class that can be mapped to an external given timer; this
will need a bit of testing and debugging most likely. Implemented into
Robot.java.